+++ modules/webcam/onvifpr.cpp
@@ -16,7 +16,7 @@ int convifpr::play(bool store){
 	m_frmevt  = event_create( false, true );
 	mycapture = cvCaptureFromFile(rtpurl);
 #ifdef WIN32
-	threadid  = _beginthreadex(NULL,0,playthread,this,0,NULL);
+	threadid  = (unsigned int)_beginthreadex(NULL,0,playthread,this,0,NULL);
 #else
 	if( pthread_create(&threadid, NULL, playthread, this) != 0) 
 		printf("pthread_create failed! \n");
