+++ modules/webcam/snaps.cpp
@@ -27,7 +27,7 @@ int csnaps::play(bool store){
 	m_playevt = event_create( false, false );
 	m_frmevt  = event_create( false, true );
 #ifdef WIN32
-	threadid  = _beginthreadex(NULL,0,playthread,this,0,NULL);
+	threadid  = (unsigned int)_beginthreadex(NULL,0,playthread,this,0,NULL);
 #else
 	if( pthread_create(&threadid, NULL, playthread, this) != 0) 
 		printf("pthread_create failed! \n");
@@ -89,7 +89,7 @@ void csnaps::work(){
 		newframe.append(buf, nrecv);
 		nsize -= nrecv;
 	}
-	setframe(&newframe[0], newframe.size());
+	setframe(&newframe[0], (int)newframe.size());
 	return;
 }
